+++ docs/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 1.1.0 (unreleased)
 ------------------
 
+- #6 Remove IMicroorganismBehavior and make Microorganism folderish
+
 
 1.0.0 (2022-06-17)
 ------------------